moves clients in 1 folder and split azure client
