+++ api/vms.py
@@ -2,9 +2,9 @@ from fastapi import APIRouter, Depends
 from fastapi.responses import JSONResponse
 
 from auth import User, get_current_user
-from azure_client import AzureClient
-from dependencies import get_azure_client, get_guacamole_client
-from guacamole_client import GuacamoleClient, GuacamoleConnectionNotFound
+from clients.azure import VMAzureClient
+from clients.guacamole import GuacamoleClient, GuacamoleConnectionNotFound
+from dependencies import get_guacamole_client, get_vm_azure_client
 
 router = APIRouter(prefix="/vms", tags=["vms"])
 
@@ -14,7 +14,7 @@ router = APIRouter(prefix="/vms", tags=["vms"])
 def delete_vm(
     project_name: str,
     current_user: User = Depends(get_current_user),
-    azure_client: AzureClient = Depends(get_azure_client),
+    azure_client: VMAzureClient = Depends(get_vm_azure_client),
     guacamole_client: GuacamoleClient = Depends(get_guacamole_client),
 ):
     """Delete a VM and its connection information on Guacamole."""
