+++ clients/azure/data.py
@@ -1,18 +1,12 @@
 import os
 import re
-from dataclasses import dataclass
 from datetime import datetime, timedelta
 from pathlib import Path
-from typing import Any, Generator, Literal, Optional
+from typing import Generator, Literal, Optional
 
 from azure.core.credentials import TokenCredential
 from azure.core.exceptions import ResourceNotFoundError
-from azure.core.polling import LROPoller
 from azure.identity import DefaultAzureCredential
-from azure.mgmt.compute import ComputeManagementClient
-from azure.mgmt.resource import ResourceManagementClient
-from azure.mgmt.resource.resources.models import DeploymentExtended
-from azure.mgmt.resource.templatespecs import TemplateSpecsClient
 from azure.mgmt.storage import StorageManagementClient
 from azure.storage.file.models import FilePermissions
 from azure.storage.file.sharedaccesssignature import FileSharedAccessSignature
@@ -24,21 +18,13 @@ from azure.storage.fileshare import (
 )
 from dotenv import load_dotenv
 from pydantic import BaseModel
-from slugify import slugify
 
 from auth import User
 
 load_dotenv()
 
+
 RunDataTypeType = Literal["processed_data", "raw_data"]
-
-
-class DeploymentNotFound(Exception):
-    pass
-
-
-class VMNotFound(Exception):
-    pass
 
 
 class RunDataNotFound(Exception):
@@ -55,21 +41,6 @@ class IncorrectDataFilePath(Exception):
         super().__init__(*args)
 
 
-@dataclass
-class AzureVMDeploymentProperties:
-    project_name: str
-    username: str
-    password: str
-    deployment_process: LROPoller[DeploymentExtended]
-
-
-@dataclass
-class AzureCaptureDeploymentProperties:
-    project_name: str
-    version: str
-    deployment_process: LROPoller[DeploymentExtended]
-
-
 class ProjectFile(BaseModel):
     name: str
     last_modified: Optional[datetime]
@@ -77,17 +48,11 @@ class ProjectFile(BaseModel):
     path: Optional[str]
 
 
-class AzureClient:
-    """Provides an API to interact with Azure services."""
-
-    def __init__(
-        self,
-    ):
-        self.resource_group_name = os.environ["AZURE_RESOURCE_GROUP_NAME"]
-        self.template_specs_name = os.environ["AZURE_TEMPLATE_SPECS_NAME"]
-        self.storage_account_name = os.environ["AZURE_STORAGE_ACCOUNT"]
-
+class StorageAzureClient:
+    def __init__(self):
         credentials = DefaultAzureCredential()
+        self.resource_group_name = os.environ["AZURE_RESOURCE_GROUP_NAME"]
+        self.storage_account_name = os.environ["AZURE_STORAGE_ACCOUNT"]
 
         storage_key = _get_storage_key(
             credentials,
@@ -95,121 +60,14 @@ class AzureClient:
             resource_group_name=self.resource_group_name,
             storage_account_name=self.storage_account_name,
         )
+
         # pylint: disable=line-too-long,consider-using-f-string
         self._storage_connection_string = "DefaultEndpointsProtocol=https;AccountName={};AccountKey={};EndpointSuffix=core.windows.net".format(
             self.storage_account_name, storage_key
         )
-
         self._file_shared_access_signature = FileSharedAccessSignature(
             account_name=self.storage_account_name, account_key=storage_key
         )
-
-        self._resource_mgmt_client = ResourceManagementClient(
-            credentials, os.environ["AZURE_SUBSCRIPTION_ID"]
-        )
-        self._compute_mgmt_client = ComputeManagementClient(
-            credentials, os.environ["AZURE_SUBSCRIPTION_ID"]
-        )
-        self._template_specs_client = TemplateSpecsClient(
-            credentials, os.environ["AZURE_SUBSCRIPTION_ID"]
-        )
-
-    def get_vm(self, project_name: str):
-        """Retrieves VM information with project name."""
-        try:
-            return self._compute_mgmt_client.virtual_machines.get(
-                resource_group_name=self.resource_group_name,
-                vm_name=_project_name_to_vm_name(project_name),
-            )
-        except ResourceNotFoundError as error:
-            raise VMNotFound from error
-
-    def delete_deployment(self, project_name: str):
-        return self._resource_mgmt_client.deployments.begin_delete(
-            resource_group_name=self.resource_group_name,
-            deployment_name=slugify(project_name),
-        )
-
-    def get_deployment_status(
-        self, project_name: str
-    ) -> Literal[
-        "NotSpecified",
-        "Accepted",
-        "Running",
-        "Ready",
-        "Creating",
-        "Created",
-        "Canceled",
-        "Failed",
-        "Succeeded",
-        "Updating",
-    ]:
-        """Retrieves VM information."""
-        try:
-            deployment = self._resource_mgmt_client.deployments.get(
-                resource_group_name=self.resource_group_name,
-                deployment_name=slugify(project_name),
-            )
-        except ResourceNotFoundError as error:
-            raise DeploymentNotFound() from error
-        return deployment.properties.provisioning_state
-
-    def deploy_vm(
-        self,
-        project_name: str,
-        vm_size: Literal[
-            "Standard_B8ms",
-            "Standard_B20ms",
-            "Standard_DS1_v2",
-        ] = None,
-    ) -> Optional[AzureVMDeploymentProperties]:
-        """Deploys a VM based on Template Specs specified
-        with AZURE_TEMPLATE_SPECS_NAME env variable.
-        In both cases where the deployment is created or it has
-        already been created before, the function returns None.
-        """
-        if self._resource_mgmt_client.deployments.check_existence(
-            resource_group_name=self.resource_group_name,
-            deployment_name=slugify(project_name),
-        ):
-            return None
-        template = self._get_latest_template_specs(
-            template_name=self.template_specs_name
-        )
-        parameters = {
-            "vmName": slugify(project_name),
-        }
-        if vm_size:
-            parameters["vmSize"] = vm_size
-        formatted_parameters = {k: {"value": v} for k, v in parameters.items()}
-        poller = self._resource_mgmt_client.deployments.begin_create_or_update(
-            resource_group_name=self.resource_group_name,
-            deployment_name=slugify(project_name),
-            parameters={
-                "properties": {
-                    "template": template,
-                    "parameters": formatted_parameters,
-                    "mode": "Incremental",
-                },
-            },
-        )
-        return AzureVMDeploymentProperties(
-            project_name=project_name,
-            username=os.environ["VM_LOGIN"],
-            password=os.environ["VM_PASSWORD"],
-            deployment_process=poller,
-        )
-
-    def delete_vm(self, project_name: str) -> Literal["Failed", "Succeeded"]:
-        try:
-            operation = self._compute_mgmt_client.virtual_machines.begin_delete(
-                resource_group_name=self.resource_group_name,
-                vm_name=_project_name_to_vm_name(project_name),
-            )
-        except ResourceNotFoundError as error:
-            raise VMNotFound from error
-        operation.result()
-        return operation.status()
 
     def get_project_documents(
         self,
@@ -241,6 +99,20 @@ class AzureClient:
         except ResourceNotFoundError as error:
             raise RunDataNotFound from error
 
+    def generate_run_data_sas_url(
+        self,
+        dir_path: str,
+        file_name: str,
+        is_admin: bool,
+    ):
+        """Generate URL with Shared Access Signature to manage run data in an
+        Azure Fileshare. Regular users can read. Admins can also write, create & delete.
+        """
+        permission = FilePermissions(
+            read=True, create=is_admin, write=is_admin, delete=is_admin
+        )
+        return self._generate_sas_url(dir_path, file_name, permission)
+
     def generate_project_documents_upload_sas_url(
         self, project_name: str, file_name: str
     ):
@@ -258,20 +130,6 @@ class AzureClient:
         generate_project_documents_upload_sas_url.
         """
         permission = FilePermissions(read=True, create=False, write=False, delete=True)
-        return self._generate_sas_url(dir_path, file_name, permission)
-
-    def generate_run_data_sas_url(
-        self,
-        dir_path: str,
-        file_name: str,
-        is_admin: bool,
-    ):
-        """Generate URL with Shared Access Signature to manage run data in an
-        Azure Fileshare. Regular users can read. Admins can also write, create & delete.
-        """
-        permission = FilePermissions(
-            read=True, create=is_admin, write=is_admin, delete=is_admin
-        )
         return self._generate_sas_url(dir_path, file_name, permission)
 
     def _generate_sas_url(
@@ -321,35 +179,6 @@ class AzureClient:
                     ],
                 )
             ]
-        )
-
-    def create_new_image_version(self, project_name: str, version: str):
-        """
-        Will use the given vm to create a new specialized image of this image and save it
-        to the image gallery with the given version
-        """
-        vm_name = _project_name_to_vm_name(project_name)
-        template = self._get_latest_template_specs(template_name="captureVMSpec")
-        parameters = {"vmName": vm_name, "version": version}
-
-        formatted_parameters = {k: {"value": v} for k, v in parameters.items()}
-
-        poller = self._resource_mgmt_client.deployments.begin_create_or_update(
-            resource_group_name=self.resource_group_name,
-            deployment_name=f"updatevmimage_{vm_name}_{version}",
-            parameters={
-                "properties": {
-                    "template": template,
-                    "parameters": formatted_parameters,
-                    "mode": "Incremental",
-                },
-            },
-        )
-
-        return AzureCaptureDeploymentProperties(
-            project_name=vm_name,
-            version=version,
-            deployment_process=poller,
         )
 
     def _list_files_recursive(
@@ -415,38 +244,25 @@ class AzureClient:
                 else:
                     yield ProjectFile.parse_obj({**file, "path": path})
 
-    def _get_latest_template_specs(self, template_name: str) -> dict[str, Any]:
-        """Get latest template specs in a python dict format."""
-        template_spec = self._template_specs_client.template_specs.get(
-            resource_group_name=self.resource_group_name,
-            template_spec_name=template_name,
-            expand="versions",
-        )
-        latest_version = sorted(template_spec.versions.keys())[-1]
-        return self._template_specs_client.template_spec_versions.get(
-            resource_group_name=self.resource_group_name,
-            template_spec_name=template_name,
-            template_spec_version=latest_version,
-        ).main_template
 
-
-def _project_name_to_vm_name(project_name: str):
-    """Returns a correct vm name (prefix added, slugified) based on a project name"""
-    # pylint: disable=consider-using-f-string
-    return "{}{}".format(os.getenv("AZURE_RESOURCE_PREFIX"), slugify(project_name))
-
-
-def wait_for_deployment_completeness(
-    poller: LROPoller[DeploymentExtended],
-) -> Optional[DeploymentExtended]:
-    deployment = poller.result()
-    if deployment.properties.provisioning_state in (
-        "Succeeded",
-        "Running",
-        "Ready",
+def validate_run_data_file_path(path: Path, current_user: User):
+    if not re.match(
+        rf"^{_get_projects_path()}\/[\w\- ]+\/runs\/[\w\- ]+\/(raw_data|processed_data)\/",
+        str(path),
     ):
-        return deployment
-    return None
+        # pylint: disable=line-too-long
+        raise IncorrectDataFilePath(
+            "path must start with {projects_path_prefix}/<project_name>/runs/<run_name>/(processed_data|raw_data)/"
+        )
+    _validate_project_file_path(path, current_user)
+
+
+def validate_project_document_file_path(path: Path, current_user: User):
+    if not re.match(rf"^{_get_projects_path()}\/[\w\- ]+\/documents", str(path)):
+        raise IncorrectDataFilePath(
+            "path must start with {projects_path_prefix}/<project_name>/documents/"
+        )
+    _validate_project_file_path(path, current_user)
 
 
 def _get_storage_key(
@@ -481,23 +297,3 @@ def _validate_project_file_path(path: Path, current_user: User):
     project_name = path_without_prefix.parts[0]
     if not current_user.has_project(project_name) and not current_user.is_admin:
         raise IncorrectDataFilePath(f"user is not part of project {project_name}")
-
-
-def validate_run_data_file_path(path: Path, current_user: User):
-    if not re.match(
-        rf"^{_get_projects_path()}\/[\w\- ]+\/runs\/[\w\- ]+\/(raw_data|processed_data)\/",
-        str(path),
-    ):
-        # pylint: disable=line-too-long
-        raise IncorrectDataFilePath(
-            "path must start with {projects_path_prefix}/<project_name>/runs/<run_name>/(processed_data|raw_data)/"
-        )
-    _validate_project_file_path(path, current_user)
-
-
-def validate_project_document_file_path(path: Path, current_user: User):
-    if not re.match(rf"^{_get_projects_path()}\/[\w\- ]+\/documents", str(path)):
-        raise IncorrectDataFilePath(
-            "path must start with {projects_path_prefix}/<project_name>/documents/"
-        )
-    _validate_project_file_path(path, current_user)
