+++ dependencies.py
@@ -1,12 +1,17 @@
 from functools import lru_cache
 
-from azure_client import AzureClient
-from guacamole_client import GuacamoleClient
+from clients.azure import StorageAzureClient, VMAzureClient
+from clients.guacamole import GuacamoleClient
 
 
 @lru_cache()
-def get_azure_client():
-    return AzureClient()
+def get_vm_azure_client():
+    return VMAzureClient()
+
+
+@lru_cache()
+def get_storage_azure_client():
+    return StorageAzureClient()
 
 
 @lru_cache()
