+++ scripts/delete_vm.py
@@ -5,8 +5,8 @@ VM & deployment won't raise if they don't exist on Azure. Guacamole connection w
 """
 import argparse
 
-from azure_client import AzureClient
-from guacamole_client import GuacamoleClient, GuacamoleConnectionNotFound
+from clients.azure import VMAzureClient
+from clients.guacamole import GuacamoleClient, GuacamoleConnectionNotFound
 
 from . import get_logger
 
@@ -20,13 +20,15 @@ def delete_vm():
     )
     args = parser.parse_args()
 
+    azure_client = VMAzureClient()
+
     logger.info("Deleting Azure VM...")
-    status = AzureClient().delete_vm(args.project_name)
+    status = azure_client.delete_vm(args.project_name)
     logger.info("Azure VM deletion operation finished with satus : %s", status)
     if status == "Failed":
         logger.error("Couldn't delete Azure VM.")
     logger.info("Deleting Azure deployment...")
-    AzureClient().delete_deployment(args.project_name)
+    azure_client.delete_deployment(args.project_name)
     logger.info("Deleting Guacamole connection...")
     try:
         GuacamoleClient().delete_connection(args.project_name)
