+++ scripts/generate_file_sas_url.py
@@ -4,7 +4,7 @@ Create a Azure VM and a Guacamole connection
 """
 import argparse
 
-from azure_client import AzureClient
+from clients.azure import StorageAzureClient
 
 
 def generate_file_sas_url():
@@ -13,7 +13,7 @@ def generate_file_sas_url():
     parser.add_argument("file_name", help="File name.")
     args = parser.parse_args()
 
-    azure_client = AzureClient()
+    azure_client = StorageAzureClient()
 
     url = azure_client.generate_run_data_sas_url(
         dir_path=args.dir_path,
