+++ tests/clients/test_azure_data_client.py
@@ -4,207 +4,33 @@ from datetime import datetime
 from unittest.mock import DEFAULT, MagicMock, patch
 
 import pytest
-from azure.core.exceptions import ResourceNotFoundError
 from pytest import MonkeyPatch
 
 from auth import Project, User
-from azure_client import (
-    AzureCaptureDeploymentProperties,
-    AzureClient,
-    AzureVMDeploymentProperties,
-    DeploymentNotFound,
+from clients.azure import StorageAzureClient
+from clients.azure.data import (
     IncorrectDataFilePath,
     ProjectFile,
-    VMNotFound,
-    _project_name_to_vm_name,
     validate_project_document_file_path,
     validate_run_data_file_path,
-    wait_for_deployment_completeness,
 )
 
 
 @pytest.fixture
 def client(monkeypatch: MonkeyPatch):
     monkeypatch.setenv("AZURE_RESOURCE_GROUP_NAME", "resource_group_name")
-    monkeypatch.setenv("AZURE_TEMPLATE_SPECS_NAME", "template_specs")
     monkeypatch.setenv("AZURE_SUBSCRIPTION_ID", "ID")
-    monkeypatch.setenv("AZURE_RESOURCE_PREFIX", "test-")
     monkeypatch.setenv("AZURE_STORAGE_ACCOUNT", "storageaccount")
-    monkeypatch.setenv("VM_LOGIN", "username")
-    monkeypatch.setenv("VM_PASSWORD", "password")
     with patch.multiple(
-        "azure_client",
-        ResourceManagementClient=DEFAULT,
-        ComputeManagementClient=DEFAULT,
-        TemplateSpecsClient=DEFAULT,
+        "clients.azure.data",
         StorageManagementClient=DEFAULT,
         FileSharedAccessSignature=DEFAULT,
     ):
-        return AzureClient()
-
-
-@patch("azure_client.AzureClient._get_latest_template_specs", dict)
-def test_deploy_exits_when_vm_exists(client: AzureClient):
-    client._resource_mgmt_client.deployments.check_existence.return_value = True
-    client.deploy_vm("vm-test")
-
-    client._resource_mgmt_client.deployments.begin_create_or_update.assert_not_called()
-
-
-@patch("azure_client.AzureClient._get_latest_template_specs", dict)
-@patch("azure_client._project_name_to_vm_name", lambda x: x)
-def test_deploys_with_proper_parameters(client: AzureClient):
-    client._resource_mgmt_client.deployments.check_existence.return_value = False
-    result = client.deploy_vm("vm-test", vm_size="Standard_B8ms")
-
-    call_args = (
-        client._resource_mgmt_client.deployments.begin_create_or_update.call_args[1]
-    )
-    assert "parameters" in call_args
-    assert "properties" in call_args["parameters"]
-    assert "template" in call_args["parameters"]["properties"]
-    assert "parameters" in call_args["parameters"]["properties"]
-    assert (
-        call_args["parameters"]["properties"]["parameters"]["vmName"]["value"]
-        == "vm-test"
-    )
-    assert (
-        call_args["parameters"]["properties"]["parameters"]["vmSize"]["value"]
-        == "Standard_B8ms"
-    )
-    assert isinstance(result, AzureVMDeploymentProperties)
-    assert result.project_name == "vm-test"
-    assert result.username == "username"
-    assert isinstance(result.password, str)
-    assert (
-        result.deployment_process
-        is client._resource_mgmt_client.deployments.begin_create_or_update.return_value
-    )
-
-
-@patch("azure_client.AzureClient._get_latest_template_specs", dict)
-@patch("azure_client._project_name_to_vm_name", lambda x: x)
-def test_create_image(client: AzureClient):
-    client._resource_mgmt_client.deployments.check_existence.return_value = False
-    result = client.create_new_image_version("vm-test", version="1.1.1")
-
-    call_args = (
-        client._resource_mgmt_client.deployments.begin_create_or_update.call_args[1]
-    )
-
-    assert "parameters" in call_args
-    assert "properties" in call_args["parameters"]
-    assert "template" in call_args["parameters"]["properties"]
-    assert "parameters" in call_args["parameters"]["properties"]
-    assert (
-        call_args["parameters"]["properties"]["parameters"]["vmName"]["value"]
-        == "vm-test"
-    )
-    assert (
-        call_args["parameters"]["properties"]["parameters"]["version"]["value"]
-        == "1.1.1"
-    )
-    assert isinstance(result, AzureCaptureDeploymentProperties)
-    assert result.project_name == "vm-test"
-    assert result.version == "1.1.1"
-    assert (
-        result.deployment_process
-        is client._resource_mgmt_client.deployments.begin_create_or_update.return_value
-    )
-
-
-def test_get_latest_template_specs(client: AzureClient):
-    client._template_specs_client.template_specs.get.return_value.versions = {
-        "1.0.0": {},
-        "1.1.1": {},
-    }
-    client._get_latest_template_specs(template_name="template_specs")
-
-    client._template_specs_client.template_spec_versions.get.assert_called_with(
-        resource_group_name="resource_group_name",
-        template_spec_name="template_specs",
-        template_spec_version="1.1.1",
-    )
-
-
-def test_get_vm_raises_if_absent(client: AzureClient):
-    client._compute_mgmt_client.virtual_machines.get.side_effect = (
-        ResourceNotFoundError()
-    )
-    with pytest.raises(VMNotFound):
-        client.get_vm("VM")
-
-
-@patch("azure_client._project_name_to_vm_name", lambda x: x)
-def test_get_vm_calls_azure_method_with_proper_args(client: AzureClient):
-    client.get_vm("VM")
-    client._compute_mgmt_client.virtual_machines.get.assert_called_with(
-        resource_group_name="resource_group_name",
-        vm_name="VM",
-    )
-
-
-def test_get_deployment_status_returns_status(client: AzureClient):
-    deployment_get_return = MagicMock(
-        properties=MagicMock(provisioning_state="Succeeded")
-    )
-    client._resource_mgmt_client.deployments.get.return_value = deployment_get_return
-    status = client.get_deployment_status("VM")
-
-    client._resource_mgmt_client.deployments.get.assert_called_with(
-        resource_group_name="resource_group_name", deployment_name="vm"
-    )
-    assert status == "Succeeded"
-
-
-def test_get_deployment_raises_if_deployment_absent(client: AzureClient):
-    client._resource_mgmt_client.deployments.get.side_effect = ResourceNotFoundError()
-
-    with pytest.raises(DeploymentNotFound):
-        client.get_deployment_status("VM")
-
-
-@pytest.mark.parametrize(
-    ("status", "is_ok"),
-    (("Succeeded", True), ("Running", True), ("Ready", True), ("Failed", False)),
-)
-def test_wait_for_deployment_completeness(status, is_ok):
-    poller = MagicMock(result=MagicMock())
-    poller.result.return_value = MagicMock(
-        properties=MagicMock(provisioning_state=status)
-    )
-    deployment = wait_for_deployment_completeness(poller)
-    if is_ok:
-        assert deployment
-    else:
-        assert not deployment
-
-
-def test_delete_vm(client: AzureClient):
-    client._resource_mgmt_client.deployments.check_existence.return_value = False
-    client.delete_vm("vm-test")
-
-    client._compute_mgmt_client.virtual_machines.begin_delete.assert_called_with(
-        resource_group_name="resource_group_name",
-        vm_name="test-vm-test",
-    )
-
-
-def test_delete_vm_raises_if_vm_absent(client: AzureClient):
-    client._compute_mgmt_client.virtual_machines.begin_delete.side_effect = (
-        ResourceNotFoundError()
-    )
-    with pytest.raises(VMNotFound):
-        client.delete_vm("vm-test")
-
-
-def test_project_name_to_vm_name(monkeypatch: MonkeyPatch):
-    monkeypatch.setenv("AZURE_RESOURCE_PREFIX", "test-")
-    assert _project_name_to_vm_name("BLABLA") == "test-blabla"
+        return StorageAzureClient()
 
 
 def test_get_project_documents_with_prefix(
-    client: AzureClient, monkeypatch: MonkeyPatch
+    client: StorageAzureClient, monkeypatch: MonkeyPatch
 ):
     monkeypatch.setenv("AZURE_STORAGE_PROJECTS_LOCATION_PREFIX", "/prefix")
     _list_files_recursive_mock = MagicMock(
@@ -227,7 +53,9 @@ def test_get_project_documents_with_prefix(
         assert len(list(files)) == 1
 
 
-def test_get_run_files_with_prefix(client: AzureClient, monkeypatch: MonkeyPatch):
+def test_get_run_files_with_prefix(
+    client: StorageAzureClient, monkeypatch: MonkeyPatch
+):
     monkeypatch.setenv("AZURE_STORAGE_PROJECTS_LOCATION_PREFIX", "/prefix")
     _list_files_recursive_mock = MagicMock(
         return_value=(
@@ -252,7 +80,7 @@ def test_get_run_files_with_prefix(client: AzureClient, monkeypatch: MonkeyPatch
 
 
 def test_generate_project_documents_sas_url(
-    client: AzureClient,
+    client: StorageAzureClient,
     monkeypatch: MonkeyPatch,
 ):
     with patch.object(
@@ -282,10 +110,10 @@ def test_generate_project_documents_sas_url(
     assert mock_kwargs["permission"].write is False
 
 
-@patch("azure_client._get_projects_path")
+@patch("clients.azure.data._get_projects_path")
 def test_generate_project_documents_upload_sas_url(
     _get_projects_path_mock: MagicMock,
-    client: AzureClient,
+    client: StorageAzureClient,
     monkeypatch: MonkeyPatch,
 ):
     with patch.object(
@@ -321,7 +149,7 @@ def test_generate_project_documents_upload_sas_url(
     (True, False),
 )
 def test_generate_run_data_sas_url(
-    client: AzureClient,
+    client: StorageAzureClient,
     monkeypatch: MonkeyPatch,
     is_admin: bool,
 ):
@@ -352,12 +180,12 @@ def test_generate_run_data_sas_url(
     assert mock_kwargs["permission"].write == is_admin
 
 
-@patch("azure_client.ShareDirectoryClient")
-@patch("azure_client.ShareFileClient")
+@patch("clients.azure.data.ShareDirectoryClient")
+@patch("clients.azure.data.ShareFileClient")
 def test_list_files_recursive_without_detailed_info(
     share_file_client: MagicMock,
     share_directory_client: MagicMock,
-    client: AzureClient,
+    client: StorageAzureClient,
     monkeypatch: MonkeyPatch,
 ):
     monkeypatch.setenv("AZURE_STORAGE_FILESHARE", "fileshare")
@@ -398,12 +226,12 @@ def test_list_files_recursive_without_detailed_info(
     assert all(file.last_modified is None for file in files_list)
 
 
-@patch("azure_client.ShareDirectoryClient")
-@patch("azure_client.ShareFileClient")
+@patch("clients.azure.data.ShareDirectoryClient")
+@patch("clients.azure.data.ShareFileClient")
 def test_list_files_recursive_with_detailed_info(
     share_file_client: MagicMock,
     share_directory_client: MagicMock,
-    client: AzureClient,
+    client: StorageAzureClient,
     monkeypatch: MonkeyPatch,
 ):
     monkeypatch.setenv("AZURE_STORAGE_FILESHARE", "fileshare")
@@ -463,7 +291,7 @@ def test_list_files_recursive_with_detailed_info(
     )
 
 
-@patch("azure_client._get_projects_path", MagicMock(return_value="projects"))
+@patch("clients.azure.data._get_projects_path", MagicMock(return_value="projects"))
 @pytest.mark.parametrize(
     ("path,is_valid"),
     (
@@ -489,7 +317,7 @@ def test_validate_run_data_file_path(path, is_valid):
     assert is_valid is not is_invalid
 
 
-@patch("azure_client._get_projects_path", MagicMock(return_value="projects"))
+@patch("clients.azure.data._get_projects_path", MagicMock(return_value="projects"))
 @pytest.mark.parametrize(
     ("path,is_valid"),
     (
